+++ app/test/e2e/dataset-flush.spec.js
@@ -16,8 +16,6 @@ describe('Upload raw data', () => {
         if (process.env.NODE_ENV !== 'test') {
             throw Error(`Running the test suite with NODE_ENV ${process.env.NODE_ENV} may result in permanent data loss. Please use NODE_ENV=test.`);
         }
-
-        nock.cleanAll();
     });
 
     it('Return 401 error if no user provided', async () => {