+++ app/test/e2e/dataset-upload.spec.js
@@ -18,8 +18,6 @@ describe('Upload raw data', () => {
             throw Error(`Running the test suite with NODE_ENV ${process.env.NODE_ENV} may result in permanent data loss. Please use NODE_ENV=test.`);
         }
 
-        nock.cleanAll();
-
         await Dataset.deleteMany({}).exec();
     });